Bump guest RAM to ~2 GiB; keep SP at KSEG0 top
RamDevice stays 2 GiB from phys 0. 0x80000000+RamSize-0x1000 overflows
uint at 2 GiB and exceeds the 512 MiB KSEG0 window, so SP uses
min(RamSize, 0x20000000) and stays 0x9FFFF000. Long alloc + very-large
objects so the 2 GiB byte[] is not a negative int. No new maps.

+++ RamDevice.cs
@@ -12,7 +12,7 @@ namespace ProcessorEmulator.Emulation
         {
             StartAddress = startAddress;
             Size = size;
-            _memory = new byte[size];
+            _memory = new byte[(long)size];
         }
 
         public uint Read32(uint offset)
